refactor: separate wire opcodes from game dispatch
