+++ src/CMakeLists.txt
@@ -20,6 +20,8 @@
 # Build the mangos shared library
 add_subdirectory(shared)
 
+add_subdirectory(proto)
+
 add_subdirectory(genrev)
 
 # Needs to link against mangos_world.lib
