+++ src/game/ChatCommands/DebugCommands.cpp
@@ -37,7 +37,7 @@
 #include "DBCStores.h"
 #include "WorldPacket.h"
 #include "Player.h"
-#include "Opcodes.h"
+#include "OpcodeTable.h"
 #include "Chat.h"
 #include "Log.h"
 #include "Unit.h"
@@ -293,7 +293,7 @@ bool ChatHandler::HandleDebugRecvOpcodeCommand(char* /*args*/)
     }
     stream.close();
 
-    DEBUG_LOG("Queued opcode %u, %s", data->GetOpcode(), data->GetOpcodeName());
+    DEBUG_LOG("Queued opcode %u, %s", data->GetOpcode(), LookupOpcodeName(data->GetOpcode()));
 
     m_session->QueuePacket(data);
 
@@ -406,7 +406,7 @@ bool ChatHandler::HandleDebugSendOpcodeCommand(char* /*args*/)
     }
     stream.close();
 
-    DEBUG_LOG("Sending opcode %u, %s", data.GetOpcode(), data.GetOpcodeName());
+    DEBUG_LOG("Sending opcode %u, %s", data.GetOpcode(), LookupOpcodeName(data.GetOpcode()));
 
     data.hexlike();
     unit->ToPlayer()->SendDirectMessage(&data);
