+++ src/proto/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(proto INTERFACE)
+target_include_directories(proto INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+target_link_libraries(proto INTERFACE shared)
